requirements.txt
streamlit
torch
transformers
google-generativeai
python-dotenv